+++ src/Geta.Optimizely.GenericLinks.Web/Program.cs
@@ -1,0 +1,11 @@
+using Geta.Optimizely.GenericLinks.Web;
+
+Host.CreateDefaultBuilder(args)
+    .ConfigureCmsDefaults()
+    .ConfigureWebHostDefaults(webBuilder =>
+    {
+        webBuilder.UseStartup<Startup>();
+        webBuilder.UseContentRoot(Path.GetFullPath("../../sub/geta-foundation-core/src/Foundation"));
+    })
+    .Build()
+    .Run();