+++ src/Geta.Optimizely.GenericLinks/Html/DefaultLinkHtmlSerializer.cs
@@ -3,7 +3,6 @@
 
 using System.Net;
 using System.Text;
-using EPiServer.Web;
 using EPiServer.Web.Routing;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,12 +12,10 @@ namespace Geta.Optimizely.GenericLinks.Html;
 
 public class DefaultLinkHtmlSerializer : ILinkHtmlSerializer
 {
-    private readonly IVirtualPathResolver _virtualPathResolver;
     private readonly IUrlResolver _urlResolver;
 
-    public DefaultLinkHtmlSerializer(IVirtualPathResolver virtualPathResolver, IUrlResolver urlResolver)
+    public DefaultLinkHtmlSerializer(IUrlResolver urlResolver)
     {
-        _virtualPathResolver = virtualPathResolver;
         _urlResolver = urlResolver;
     }
 
@@ -91,7 +88,7 @@ public class DefaultLinkHtmlSerializer : ILinkHtmlSerializer
 
         if (mode == StringMode.ViewMode)
         {
-            href = _virtualPathResolver.ToAbsoluteOrSame(link.Href);
+            href = VirtualPathHelper.ToAbsoluteOrSame(link.Href);
         }
         else
         {